Spacemit krill (#771)
* cercando di modificare krill

* entriamo all'otto

* boh

* 26.3.8

* ultimi progressi

* rinforzando mountfs

* cercando di sistemare spacemit

* dovrebbe aver scritto anche il bootloader uboot in partition

* pnpm update

* 26.3.10

* active

* ri-funziona

* 26.3.11

* v26.3.11

* che casino

* just pnpm update

+++ src/krill/classes/sequence.d/mount_fs.ts
@@ -8,10 +8,9 @@
  */
 
 import fs from 'node:fs'
-
-import Utils from '../../../classes/utils.js'
+import os from 'node:os'
 import { exec } from '../../../lib/utils.js'
-import { InstallationMode, SwapChoice } from '../krill_enums.js'
+import { SwapChoice } from '../krill_enums.js'
 import Sequence from '../sequence.js'
 
 /**
@@ -24,6 +23,9 @@ export async function mountFs(this: Sequence): Promise<boolean> {
 
   // root
   await exec(`mount -t ${this.devices.root.fsType} ${this.devices.root.name} ${this.installTarget}${this.devices.root.mountPoint} ${this.toNull}`, this.echo)
+  if ((await exec(`mountpoint -q ${this.installTarget}${this.devices.root.mountPoint}`)).code !== 0) {
+    throw new Error(`[Krill ENOSPC Prevention] Failed to mount ${this.devices.root.name} on ${this.installTarget}${this.devices.root.mountPoint}`)
+  }
   await exec(`tune2fs -c 0 -i 0 ${this.devices.root.name} ${this.toNull}`, this.echo)
   await exec(`rm -rf ${this.installTarget}/lost+found ${this.toNull}`, this.echo)
 
@@ -31,22 +33,33 @@ export async function mountFs(this: Sequence): Promise<boolean> {
   if (this.devices.boot.name !== 'none') {
     await exec(`mkdir ${this.installTarget}/boot -p ${this.toNull}`, this.echo)
     await exec(`mount -t ${this.devices.boot.fsType} ${this.devices.boot.name} ${this.installTarget}${this.devices.boot.mountPoint} ${this.toNull}`, this.echo)
+    if ((await exec(`mountpoint -q ${this.installTarget}${this.devices.boot.mountPoint}`)).code !== 0) {
+      throw new Error(`Failed to mount boot partition ${this.devices.boot.name}`)
+    }
     await exec(`tune2fs -c 0 -i 0 ${this.devices.boot.name} ${this.toNull}`, this.echo)
   }
 
   // data
-  if (this.devices.data.name !== 'none') {
-    await exec(`mkdir ${this.installTarget}${this.devices.data.mountPoint} -p ${this.toNull}`, this.echo)
-    await exec(`mount -t ${this.devices.data.fsType} ${this.devices.data.name} ${this.installTarget}${this.devices.data.mountPoint} ${this.toNull}`, this.echo)
-    await exec(`tune2fs -c 0 -i 0 ${this.devices.data.name} ${this.toNull}`, this.echo)
-  }
+  if (os.arch() !== 'riscv64') {
+    if (this.devices.data.name !== 'none') {
+      await exec(`mkdir ${this.installTarget}${this.devices.data.mountPoint} -p ${this.toNull}`, this.echo)
+      await exec(`mount -t ${this.devices.data.fsType} ${this.devices.data.name} ${this.installTarget}${this.devices.data.mountPoint} ${this.toNull}`, this.echo)
+      if ((await exec(`mountpoint -q ${this.installTarget}${this.devices.data.mountPoint}`)).code !== 0) {
+        throw new Error(`Failed to mount data partition ${this.devices.data.name}`)
+      }
+      await exec(`tune2fs -c 0 -i 0 ${this.devices.data.name} ${this.toNull}`, this.echo)
+    }
 
-  // efi
-  if (this.efi && !fs.existsSync(this.installTarget + this.devices.efi.mountPoint)) {
-    await exec(`mkdir ${this.installTarget}${this.devices.efi.mountPoint} -p ${this.toNull}`, this.echo)
+    // efi
+    if (this.efi && !fs.existsSync(this.installTarget + this.devices.efi.mountPoint)) {
+      await exec(`mkdir ${this.installTarget}${this.devices.efi.mountPoint} -p ${this.toNull}`, this.echo)
 
-    // utilizzare vfat per evitare errori
-    await exec(`mount -t vfat ${this.devices.efi.name} ${this.installTarget}${this.devices.efi.mountPoint} ${this.toNull}`, this.echo)
+      // utilizzare vfat per evitare errori
+      await exec(`mount -t vfat ${this.devices.efi.name} ${this.installTarget}${this.devices.efi.mountPoint} ${this.toNull}`, this.echo)
+      if ((await exec(`mountpoint -q ${this.installTarget}${this.devices.efi.mountPoint}`)).code !== 0) {
+        throw new Error(`Failed to mount efi partition ${this.devices.efi.name}`)
+      }
+    }
   }
 
   // swap file if we need
@@ -63,14 +76,16 @@ export async function mountFs(this: Sequence): Promise<boolean> {
  * umountFs
  */
 export async function umountFs(this: Sequence): Promise<boolean> {
-  // efi
-  if (this.efi) {
-    await this.umount(this.devices.efi.name)
-  }
+  if (os.arch() !== 'riscv64') {
+    // efi
+    if (this.efi) {
+      await this.umount(this.devices.efi.name)
+    }
 
-  // data
-  if (this.devices.data.name !== 'none') {
-    await this.umount(this.devices.data.name)
+    // data
+    if (this.devices.data.name !== 'none') {
+      await this.umount(this.devices.data.name)
+    }
   }
 
   // boot
